class CrowdMap implemented
